# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eunomia API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "paths": {
...
    "/api/export": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Download the personal vault as JSON",
        "operationId": "export_data",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 `application/problem+json`, as built in `error.rs`.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable dotted error code, see `docs/errors.md`."
          },
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The HTTP reason phrase."
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "eunomia_session"
      }
    }
  }
}